Implement field_at[_mut] for Struct DynamicVariant (#20424)
# Objective

This implements the field_at methods for Struct DynamicEnum Variants
instead of only the Tuple variants. This brings it inline with
DynamicStructs and the docs for the Enum trait.

Fixes #20402 

## Solution

Instead of only matching on the Tuple variant, it matches on the Struct
variant as well

## Testing

Let me know if there's some place where unit tests should be added. I
couldn't see any existing ones, but I can add them if needed.