Use the hierachy plugin to propagate `ComputedNodeContext` (#20432)
# Objective

From tim-blackbird on discord:
```
Spawning an entity (without Node) as a child of a parent Node and then adding a Node component to that child a frame later causes the child not to appear (same for Text as in the given example).
Setting the system to run at the first frame (FrameCount(0)) makes it function as expected 
```
```rust
use bevy::{diagnostic::FrameCount, prelude::*};

fn main() {
    App::new()
    .add_plugins(DefaultPlugins)
    .add_systems(Startup, setup)
    .add_systems(Update, insert_text.run_if(resource_equals(FrameCount(1))))
    .run();
}

fn setup(mut commands: Commands) {
    commands.spawn((Node::default(), children![Text::new("Gamer"), Marker]));
    commands.spawn(Camera3d::default());
}

fn insert_text(entity: Single<Entity, With<Marker>>, mut commands: Commands) {
    commands.entity(*entity).insert(Text::new("remaG"));
}

#[derive(Component)]
struct Marker;
```

## Solution

Use the hierarchy plugin to propagate `ComputedNodeContext`.

The problem was that the `ComputedNodeContext` isn't updated immediately
and so the text isn't rendered because the extraction function can't
find a render target.

## Testing

Run the example above.

---------

Co-authored-by: Tim Blackbird <justthecooldude@gmail.com>