# Objective
- We are currently following a convention that I see repeatedly across the
crates.io ecosystem.

We provide a wrapper App::new() as a proxy for App::default()

This comes from a RUST convention that default is the preferred constructor
(when no additional parameters are required).

As a matter of library design we provide App::new only for the flexibility of
permitting the game engine developer's whim.

This PR takes the line that when documenting the codebase, when showcasing examples
or when testing the codebase we should used the preferred constructor as a way
of nudging the reader down that path.

## Solution

- Global Replace of App::new() for App::default()
- Visual inspection of PR to ensure only pertinent changes are made.

## Testing

 - Lots of tests have been updated, no change in functionality