bevy_window: Split `RawHandleWrapper` out into `Display` wrapper
Long ago `raw-window-handle` and `winit` split out a `RawDisplayHandle`
type and respective traits for dealing with the connection - or at the
very least type - of a compositor, and typically implement this for a
`Window` directly.

`wgpu` and `bevy` seem to have caught on to the latter and folded
the two traits/types together because `Window` provides it, but miss
a critical goal here: that `(Raw)DisplayHandle` is important for
initialization of certain graphics APIs. In the case of Wayland all
resources are unique per connection, and in general for others
it's important to distinguish between Wayland and X11 if `winit` chose
one over the other, even if both are available; currently that's just
guesswork inside `wgpu`.

Newer APIs like Vulkan don't suffer from this, but older graphics APIs
like OpenGL and specifically the EGL backend (or GLX for X11) set up
their entire state based on the compositor connection (alternatives are
available) even if it's ultimately "only" important for the surface that
is going to be rendered into.

Note that I haven't yet checked all the safety constraints carefully
in this PR; some existing messages seem flawed but I need to perform a
clean audit from scratch to denote what limitations should apply to the
newly proposed `RawDisplayHandleWrapper` as well.