Make shadows respect render layers (#20327)
Adopted from and closes #18747. Original work by @Adriigbs.

Shadows and render layers are a bit complicated. Ideally, this filtering
would be done at a higher level, but the camera entity just holds a list
of references to the light entities for that view. Those light entities
then compute what entities are visible from their perspective. We don't
actually get a chance to filter camera and light visible entity until we
queue the shadow pass, which is where this PR chooses to filter out
entities that aren't visible in a camera from casting shadows.

The alternative would be to store a per camera per light list of
entities, which would greatly increase the amount of memory required for
visible entities in multi-camera setups.

---------

Signed-off-by: Adrian Graur <adrian.graur@tecnico.ulisboa.pt>
Co-authored-by: Adrian Graur <adrian.graur@tecnico.ulisboa.pt>
Co-authored-by: Adrian Graur <adri.graur16@gmail.com>