Return an error from `Assets::insert` instead of panicking. (#20439)
# Objective

- Inserting into a dropped AssetId would panic. There is no API to check
if the asset ID is still valid, so you can't avoid this case.

For a concrete example, imagine a user uses `Assets::reserve_handle` to
get a handle, and then passes its AssetId into an async task. While the
async task is running, the user drops the handle, which removes the
entry from `Assets`. Later when the async task returns, it tries to
insert the asset into that `AssetId`. This previously caused a panic.
[Based on a true
story](https://discord.com/channels/691052431525675048/749332104487108618/1402511976898498590).

## Solution

- Make `Assets::insert` return an error when inserting into a dropped
`AssetId`.

## Testing

- Added a test specifically for this case.

---------

Co-authored-by: Greeble <166992735+greeble-dev@users.noreply.github.com>