Rename the `Allows` query filter to `Allow` (#20541)
This is the bikeshed of all time, but I needed to bring this up before
it became a breaking change.

In my opinion, `Allow<Disabled>` sounds better than `Allows<Disabled>`.

I don't think there's a 1-to-1 comparison in other query keywords yet,
but there's the hypothetical `Expect` in [the discussion on
#19489](https://github.com/bevyengine/bevy/pull/19489#issuecomment-3132926829),
as opposed to `Expects`.

`Allows` hasn't been in a release yet, so it's free to change until 0.17
comes out.