Improve TextSpan documentation (#20484)
# Objective

The example in the documentation of `TextSpan` is wrong. It correctly
states that `Text` or `Text2d` must be present but won't use either
leading to no rendering. It also states that `TextLayout`, `TextFont`
and `TextColor` of the first span will all set default values for the
children, which is not true. `TextLayout` determines the layout of the
block but each node has its own `TextFont` and `TextColor` which are
both required components of `TextSpan`.

The type description is not clear about `Text` or `Text2d` being
mandatory.

## Solution

- Fix the documentation