Improve error message for failed entity commands (#20544)
# Objective

- If an entity command is sent targeting a despawned entity, that
command will error, and then panic by default.
- This occurs even if the entity exists at the time the command was
sent.
- This can cause surprising panics when working with observers that
depend on each other, especially when using entity events that cause
entities to be despawned.
- Closes #19623.

## Solution

As outlined in #19623, this behavior isn't exactly a bug, although it is
surprising.

When you attempt to trigger an observer for an entity that doesn't
exist, Bevy should give you some indication that something has gone
wrong!

The standard mechanism we have for doing so is to return an error in the
command, which is passed to the default error handler. By default, this
panics, reproducing the problem reported in #19623.

Unfortunately, the error reporting was so bad that neither the OP nor
the contributors could quickly isolate the problem and tell them what
they should be doing differently.

The actual solution is both standard to the rest of Bevy and quite
simple: just queue the command in a way that makes the error not panic.
We should explain that!

I've added commentary to that effect on both
`EntityCommands::trigger/trigger_targets` and the error message
returned.

## Testing

I've added a new test demonstrating the fix for this failure mode.
Swapping that test to plain `.trigger` fixes the problem.

---------

Co-authored-by: JaySpruce <jsprucebruce@gmail.com>