Atmosphere generated environment map lighting (#20529)
# Objective

- Add generated environment lighting to the atmosphere for reflections
and diffuse
- Tracking issue: #20374

## Solution

- created a new atmosphere node type called environment
- up-sampling the existing sky-view lookup texture and tied to the
current view only. this atmosphere cube map pass has negligible
performance impact, however the filtering pipeline does have a small
performance impact to the example.
- using the new filtering pipeline, creates mip chain for different
roughness levels as well as diffuse

## Testing

- ran the atmosphere example

---

## Showcase

<img width="1281" height="752" alt="Screenshot 2025-08-12 at 12 19
08 PM"
src="https://github.com/user-attachments/assets/8f00ff25-5f48-4c51-b67e-abcbf421abc4"
/>

```rs
commands.spawn((
    Camera3d::default(),
    // ...
    // Renders the atmosphere to the environment map from the perspective of the camera
    AtmosphereEnvironmentMapLight::default(),
));
```

## Limitations, out of scope

- The generation does not support light probes (yet). This allows to
render the atmosphere as a light probe from any "location" within the
scene and within the overall atmosphere.
- therefore we assume that the relative scale of the scene to the
atmosphere are orders of magnitude different , this is a safe assumption
now with the current impl not officially supporting space views, yet
- the PBR directional light is still unaffected by the atmosphere (not
tinted by it) out of scope for this PR
- unrelated issue to this PR: small black pixels around the fully
reflective sphere. narrowed it down, this is a problem in the
inscattering calculation based on the depth value in the render sky
shader. however the addition of the env map light makes this problem
more "apparent". to be addressed separately.
- past work staged for further PRs, up next:
https://github.com/mate-h/bevy/pull/10