Diegetic UI example (#19199)
https://github.com/user-attachments/assets/e7b428b9-4705-4d0a-8322-c7a80994b1a4

This demonstrates general-purpose world-space picking. This is not
limited to bevy_ui, it works with any render target and any picking
backend, e.g. this could be a viewport/portal in 3d space, it could be a
custom UI system, bevy_egui, sprites, etc, etc.

Because this uses UV coordinates, it also works with anything that can
be mapped to a surface, so you can display UIs on curved surfaces, etc.


https://github.com/user-attachments/assets/2926a5a1-16c9-4179-85fa-83a69fd55a96

### Discussion

- This illustrates some of the improvements I think we need to make to
picking - the picking ID should be removed in favor of a pointer kind -
you might have multiple mouse pointers on multiple surfaces, with
different unique IDs. It's probably worth revisiting if we can use
Entity as the ID instead of a custom one. When I first wrote this, I ran
into some unexpected issues that caused with pointers being despawned
and getting a new ID. We will just need to be careful of edge cases
there.

---------

Co-authored-by: IceSentry <IceSentry@users.noreply.github.com>
Co-authored-by: Alice Cecile <alice.i.cecile@gmail.com>