debounce asset events for scene reloading (#20430)
# Objective

- examples `many_foxes` and `morph_targets` have stopped working since
#18358
- any case that load several parts of a complex asset and a scene, and
do setup on asset load will fail
- Fixes #20383

## Solution

- Debounce scene asset events so that we ignore modified event happening
too quickly after each others

## Testing

- run examples `many_foxes` or `morph_targets`

## Alternative

I *don't* think this is a good long term fix, but this should be fixed
in the 0.17 and I think alternatives are worse / too complex.

Alternatives are:
- Revert #18358. Normal scene loading is more important than hot
reloading
- Implement partial asset loading in the asset server and the gltf
loader so that when we load `file.gltf#Animation0`, only the animation
data is loaded. This would be a very good change, but too big to do for
the 0.17
- Have the asset server load parent assets only once, even when loading
subassets. This would be a good change and less complex, but I think
worse than the previous idea and kinda not compatible with it

---------

Co-authored-by: Alice Cecile <alice.i.cecile@gmail.com>