Rebrand glTF coordinate conversion to an alternative strategy that is biased towards glTF models (#20131)
# Objective

- Fixes https://github.com/bevyengine/bevy/issues/20121
- Reverts https://github.com/bevyengine/bevy/pull/19816

Per long discussion with @superdump (see
https://github.com/bevyengine/bevy/issues/19686#issuecomment-3069180465),
the new coordinate loading system is still wrong, just now biased in
favor of having correct forward semantics for models instead of cameras.
Which is still what I need, but it should not be the new default, as we
may well change `Transform::forward` instead.

## Solution

Since changing forward semantics is a bigger change, let's rebrand the
coordinate conversion code and keep it opt-in so that users that don't
use glTF cameras get correct semantics in the meantime.

## Testing

- Ran 3D examples and Blender test scenes.

---------

Co-authored-by: Robert Swain <robert.swain@gmail.com>