remove panic in TextureFormatPixelInfo::pixel_info (#20574)
# Objective

fixes #20365 

## Solution

TextureFormatPixelInfo::pixel_info now return a Result instead of panic

maybe a custom error specific to this case is needed, to not have the
other variants of `TextureAcessError`

## Testing

ran my game on it, using mesh, sprite, ui everything seems good.

**BUT** its my first time contributing this type of code, i dont know
much about rendering and its integration in bevy. **REVIEW THIS LIKE I
DON'T KNOW ANYTHING**. especially i dont know if there is some unwanted
consequence of these changes in other places.

what remains, is maybe we should add some `warn!` ? at least in picking.
Currently in sprite picking backend if the pixel data couldn't be
accessed (to determine if the alpha is greater than the threshold), the
entity is skipped. maybe a `warn!` here at least ?

(imo it shouldn't be skipped, instead be taken as if it was valid: id
rather have a sprite where the whole size of it is picked than having it
not working at all. maybe a SpritePickingSettings per entity ?)