set uv_transform in pbr_input_from_standard_material
Previously, even though the uv_transform was being accessed correctly from the material bindings; that information was not being passed on in the constructured pbr_input from pbr_input_from_standard_material.
This led to a situation where the StandardMaterial textures would properly have a uv_transform applied, but the pbr_input.material.uv_transform value was the default instead of the material's.
In turn this leads to any textures in a material extension using the wrong uvs, unless the value was accessed from the bindings directly instead of the pbr_input struct.