Implement ```VisitAssetDependencies``` for Arrays (#20805)
# Objective

- Since ```Vec<Handle<A>>``` and ```Vec<UntypedHandle>``` implement
```VisitAssetDependencies``` users might expect ```[Handle<A>; N]``` and
```[UntypedHandle; N]``` to also implement ```VisitAssetDependencies```,
but this is currently not the case.
- The ```#[dependency]``` attribute from the ```Asset``` derive macro
should work with an implementation.

## Solution

Implement ```VisitAssetDependencies``` for ```[Handle<A>; N]``` and
```[UntypedHandle; N]```. The implementations are based on the
corresponding ones for Vec.

## Testing

A test for compatibility with the derive macros was added for
```[Handle<A>; N]``` and ```[UntypedHandle; N]```.

---

## Showcase

Before:
```rust
// Can't use the Asset macro with a custom VisitAssetDependencies impl 
#[derive(TypePath)]
struct MyAsset {
    images: [Handle<Image>; 5],
}

impl VisitAssetDependencies for MyAsset {
    fn visit_dependencies(&self, visit: &mut impl FnMut(UntypedAssetId)) {
        for image in &self.images {
            visit(image.id().untyped());
        }
    }
}

impl Asset for MyAsset {}
```
After:
```rust
#[derive(Asset, TypePath)]
struct MyAsset {
    #[dependency]
    images: [Handle<Image>; 5],
}
```