Impl ```AsAssetId``` for ```Sprite``` and ```ImageNode``` (#20841)
# Objective

- Both ```Sprite``` and ```ImageNode``` have a 'canonical' handle, but
neither struct currently implements ```AsAssetId```. Hence they do not
work with the ```AssetChanged<A>``` filter. This is unfortunate, as the
ability to do some logic when the ```Image``` underlying a ```Sprite```
or ```ImageNode``` changes is likely useful.
- Address #20635

## Solution

- Implement ```AsAssetId``` for ```Sprite``` and ```ImageNode```.
```Asset = Image``` for both.

## Reservations
- Both ```Sprite``` and ```ImageNode``` have a ```texture_atlas:
Option<TextureAtlas>``` field. ```TextureAtlas``` contains a
```Handle<TextureAtlasLayout>```, and it is feasible to imagine a user
wanting to query for changes regarding this asset. Since ```AsAssetId```
only works for a single asset type, the choice must be made as to
support either ```Image``` or ```TextureAtlas``` , or possibly avoiding
an implementation for now. That said, ```Image``` is the most obvious
choice.

---

## Showcase
Now users can filter for ```Sprite``` components whose underlying
```Image``` asset has changed.

```rust
fn query_sprites_with_changed_images(query: Query<&Sprite, AssetChanged<Sprite>>) {
    for sprite in &query {
        println!("some logic with the changed image");
    }
}
```